Jhingoor/pyproject.toml
[project]
name = "jhingoor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.0.1,<5",
    "aiogram>=3.26.0",
    "asyncpg>=0.29.0",
    "cryptography>=42.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "google-auth>=2.29.0",
    "google-genai>=1.68.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow==10.3.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-telegram-bot==21.1",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
]
